Create Course Registration.java
There is a group of  friends who wish to enroll in a course together. The course has a maximum capacity of  students that can register for it. If there are  other students who have already enrolled in the course, determine if it will still be possible for all the  friends to do so or not.