Create The Cooler Dilemma 1 .java
The summer is at its peak in Chefland. Chef is planning to purchase a water cooler to keep his room cool. He has two options available:
Rent a cooler at the cost of  coins per month.
Purchase a cooler for  coins.
Given that the summer season will last for  months in Chefland, help Chef in finding whether he should rent the cooler or not.
Chef rents the cooler only if the cost of renting the cooler is strictly less than the cost of purchasing it. Otherwise, he purchases the cooler.
Print  if Chef should rent the cooler, otherwise print .